// src/types/stripeErrorCodes.ts
export enum StripeErrorCodes {
   ACCOUNT_COUNTRY_INVALID_ADDRESS = 'account-country-invalid-address',
   ACCOUNT_INFORMATION_MISMATCH = 'account-information-mismatch',
   ACCOUNT_INVALID = 'account_invalid',
   AMOUNT_TOO_LARGE = 'amount_too_large',
   AMOUNT_TOO_SMALL = 'amount_too_small',
   API_KEY_EXPIRED = 'api_key_expired',
   BALANCE_INSUFFICIENT = 'balance_insufficient',
   BANK_ACCOUNT_BAD_ROUTING_NUMBERS = 'bank_account_bad_routing_numbers'
};